MASTER: Implemented randomized armor selections
rando_armor() picks a style, randomizes choices based on the style, and returns a tuple containing (arms, legs, torso, style).  Style is returned to allow further decision choices if, say, the style is Chaotic.  Maybe, if the armor style is Chaotic, maybe randomize a Chaotic color scheme to go with it.  Regardless, it's implemented and appears to be fully functional. -Hark