COLOR: Implemented brightness() method of Color
Determination of brightness is based on a formula with saturation and value as variables.  This determination is rather simplistic in the larger more professional world of "colors".  I kinda 'eyeballed' the weights until I found a good balance.  Funny thing is, I can't think of a color style that will draw from the brightness member.  (Maybe "Emo"?)  Regardless, past Hark added it to the Color class so current Hark is implementing it.  Past Hark has been burned too many times by not listening to past past Hark.  -Hark