COLOR: Color Palette classes now functional
So much has happened since the last commit(?).  I was receiving odd behavior in that the child classes appeared to be sharing the same member variables.  This is not C++.  Apparently, it's a little known, yet documented (technically), feature of the language.  The ultimate solution was to empty the list of colors in each child classes __init__() function to start from scratch.  Now, each child class's list of color objects appears to be intentionally unique (exception: Main Armor Palette == Weapon Color Palette) and are no longer building on each other.

SIDE NOTE: print_color_object() was moved here from xrando.py.  It's a better fit.

-Hark