COLOR: Wrote ColorPalette.is_it_brown()
Use this method to build a dynamic list of brown colors for a particular palette.  First, validate this function is actually able to pick out brown from other colors.  If it works, use a similar method to add the brown colors to a list of earthy colors.  That way, the Earthy color style can randomly select from those colors.  -Hark