MASTER:  Small changes here
Added nation to printed stated in character's background.  Instead of just "Belfast", it will print "Belfast Ireland" for clarity.  Also, there was an odd error while reading firstnames.  The list of "names read" appeared to be of length 0.  I included a debug statement to print some information if it ever happens again.  -Hark